[refactor] NJT-40 page.tsx肥大化によりモジュール分割(カスタムフック作成など)
